+++ nucleus/metrics/polygon_utils.py
@@ -8,9 +8,9 @@ from scipy.optimize import linear_sum_assignment
 from nucleus.annotation import BoxAnnotation, PolygonAnnotation
 from nucleus.prediction import BoxPrediction, PolygonPrediction
 
-from .base import MetricResult
+from .base import ScalarResult
 from .errors import PolygonAnnotationTypeError
-from .geometry import GeometryPoint, GeometryPolygon, polygon_intersection_area
+from .geometry import GeometryPolygon, polygon_intersection_area
 
 BoxOrPolygonPrediction = TypeVar(
     "BoxOrPolygonPrediction", BoxPrediction, PolygonPrediction
@@ -35,19 +35,12 @@ def polygon_annotation_to_geometry(
         xmax = annotation.x + annotation.width / 2
         ymin = annotation.y - annotation.height / 2
         ymax = annotation.y + annotation.height / 2
-        points = [
-            GeometryPoint((xmin, ymin)),
-            GeometryPoint((xmax, ymin)),
-            GeometryPoint((xmax, ymax)),
-            GeometryPoint((xmin, ymax)),
-        ]
-        return GeometryPolygon(points)
+        points = [(xmin, ymin), (xmax, ymin), (xmax, ymax), (xmin, ymax)]
+        return GeometryPolygon(points=points, is_rectangle=True)
     elif isinstance(annotation, PolygonAnnotation):
         return GeometryPolygon(
-            [
-                GeometryPoint((point.x, point.y))
-                for point in annotation.vertices
-            ]
+            points=[(point.x, point.y) for point in annotation.vertices],
+            is_rectangle=False,
         )
     else:
         raise PolygonAnnotationTypeError()
@@ -282,7 +275,7 @@ def label_match_wrapper(metric_fn):
         *args,
         enforce_label_match: bool = False,
         **kwargs,
-    ) -> MetricResult:
+    ) -> ScalarResult:
         # Simply return the metric if we are not enforcing label matches.
         if not enforce_label_match:
             return metric_fn(annotations, predictions, *args, **kwargs)
@@ -298,6 +291,9 @@ def label_match_wrapper(metric_fn):
                 binned_annotations, binned_predictions, *args, **kwargs
             )
             metric_results.append(metric_result)
-        return MetricResult.aggregate(metric_results)
+        assert all(
+            isinstance(r, ScalarResult) for r in metric_results
+        ), "Expected every result to be a ScalarResult"
+        return ScalarResult.aggregate(metric_results)
 
     return wrapper